Kinetica.jl API - ASE Interface

ASE Calculator Builders

Kinetica.EMTBuilder — Type

EMTBuilder()

Builder for ASE EMT calculator.

Not accurate enough to be used for any kinetics, mostly just here for very quick tests.

(builder::EMTBuilder)(dir::String, mult::Int, chg::Int[, kwargs...])

Constructs an EMT calculator for ASE energy/force evaluation.

Kinetica.NWChemDFTBuilder — Type

NWChemDFTBuilder([, command::String="nwchem PREFIX.nwi > PREFIX.nwo", xc::String="becke97",
                 basis::Union{String, Dict{String, String}}="3-21G", maxiter::Int=50,
                 convergence::String="", adft::Bool=true, memory::String="1024 mb"])

Builder for NWChem-driven DFT calculator.

Contains basic functionality for on-the-fly DFT force/energy evaluations using some of the parameters available through ASE's interface. This is by no means an exhaustively customisable calculator builder, but could be used as the starting point for a more detailed builder if required.

(builder::NWChemDFTBuilder)(dir::String, mult::Int, chg::Int[, kwargs...])

Constructs a NWChem DFT calculator for ASE energy/force evaluation.

Kinetica.FHIAimsBuilder — Type

FHIAimsBuilder([, command::String="aims.x", xc::String="pbe",
               species_dir::String="./species_defaults/defaults_2020/tight",
               maxiter::Int=1000, sc_init_iter::Int=1001, dispersion::String="",
               sc_accuracy_rho::Union{Nothing, Float64}=nothing,
               sc_accuracy_forces::Union{Nothing, Float64}=nothing,
               sc_accuracy_etot::Union{Nothing, Float64}=nothing,
               sc_accuracy_eev::Union{Nothing, Float64}=nothing])

Builder for FHI Aims-driven DFT calculator.

Contains basic functionality for on-the-fly DFT force/energy evaluations using some of the parameters available through ASE's interface. This is by no means an exhaustively customisable calculator builder, but could be used as the starting point for a more detailed builder if required.

(builder::FHIAimsBuilder)(dir::String, mult::Int, chg::Int[, kwargs...])

Constructs a FHI Aims DFT calculator for ASE energy/force evaluation.