Kinetica.jl API - ASE Interface

Utilities

Conversion

Kinetica.frame_to_atoms — Function

frame_to_atoms(frame[, charges=nothing, magmoms=nothing])

Converts an ExtXYZ frame to an ASE Atoms object.

If the optional argument charges is provided with an integer array of formal charges on the atoms in frame, these will be applied to the resulting Atoms object through Atoms.set_initial_charges().

If the optional argument magmoms is provided with an array of initial magnetic moments for the atoms in frame, these will be applied to the resulting Atoms object through Atoms.set_initial_magnetic_moments().

Kinetica.atoms_to_frame — Function

atoms_to_frame(atoms[, ase_energy=nothing, inertias=nothing])

Converts an ASE Atoms object to an ExtXYZ frame.

Optionally accepts values ase_energy andinertiasthat can fill the "energy_ASE" and "inertias" keys of the resulting frame's "info" Dict.ase_energyshould be a Julia float in eV andinertiasshould be aVector{Float64}` for proper compatibility.

Kinetica.imaginary_ve_tol — Function

imaginary_ve_tol(imaginary_freq_tol)

Converts a tolerance value for imaginary frequencies into a tolerance value for imaginary vibrational energies.

I/O

Kinetica.save_asecalc — Function

save_asecalc(calc::ASENEBCalculator, saveto::String)

Saves the state of an ASENEBCalculator to a BSON file.

Deconstructs calc into a dictionary tree suitable for saving as BSON, such that it can be reconstructed with load_asecalc().

Saves everything except for the calculator builder calc.calc_builder, as this has an undefined structure that can contain Python bindings.

Kinetica.load_asecalc — Function

load_asecalc(calcfile::String[, calc_builder=nothing])

Loads the state of an ASENEBCalculator from a BSON file.

Reconstructs an ASENEBCalculator with the exception of its calc_builder field, which can optionally be provided to this function or otherwise set afterwards.

Kinetica.verify_sd — Function

verify_sd(calc_sd::SpeciesData, real_sd::SpeciesData)

Checks that calc_sd contains a subset of the data in real_sd.

Provided less or as many species are present in calc_sd as are in real_sd, ensures that each species numerically indexed in the former SpeciesData has a string representation which matches its numerical counterpart in real_sd.

If either assumption is invalid, throws an error since calc_sd cannot be extended to become real_sd.

Kinetica.verify_rd — Function

verify_rd(calc_rd::RxData, real_rd::RxData)

Checks that calc_rd contains a subset of the data in real_rd.

Provided less or as many species are present in calc_rd as are in real_rd, ensures that each reaction numerically indexed in the former RxData has a reaction hash which matches its numerical counterpart in real_rd.

If either assumption is invalid, throws an error since calc_rd cannot be extended to become real_rd.

Kinetica.save_optgeom — Function

save_optgeom(frame::Dict{String, Any}, sym, geom, saveto::String)

Saves an optimised frame geometry to a BSON file.

Also includes symmetry number sym and geometry identifier geom.

Kinetica.load_optgeom — Function

load_optgeom(savefile::String)

Loads an optimised geometry from a BSON file.

Kinetica.save_endpoints — Function

save_endpoints(reacsys::Dict{String, Any}, prodsys::Dict{String, Any}, saveto::String)

Saves reaction endpoint systems reacsys and prodsys to a BSON file.

Kinetica.load_endpoints — Function

load_endpoints(savefile::String)

Loads reaction endpoint systems from a BSON file.

Kinetica.save_tsdata — Function

save_tsdata(ts::Dict{String, Any}, conv, mult, sym, geom, chg, saveto::String)

Saves transition state data to a BSON file.

Aside from the ExtXYZ frame ts, saves the convergence status conv, spin multiplicity mult, charge chg, symmetry number sym and geometry identifier geom.

Kinetica.load_tsdata — Function

load_tsdata(savefile::String)

Loads transition state data from a BSON file.

Kinetica.save_vibdata — Function

save_vibdata(sd::SpeciesData, sids, ts_cache::Dict{Symbol, Any}, rid, saveto::String)

Saves vibrational analysis data to a BSON file.

Given species in sd indexed by sids and a transition state in ts_cache indexed by rid, saves calculated vibrational energies.

Kinetica.load_vibdata — Function

load_vibdata(savefile::String)

Loads vibrational analysis data from a BSON file.