Kinetica.jl API

Conditions

ConditionSet

Kinetica.ConditionSet — Type

Container for all conditions in a kinetic simulation.

Conditions can be static or variable, and variable conditions can be gradient-based or directly usable.

Contains fields for:

Symbolic representation of conditions (symbols)

Condition profile for each symbol (profiles)

Whether discrete rate constant updates are enabled for the conditions in this condition set (discrete_updates)

Discrete rate constant update timestep, is nothing if discrete_updates = false (ts_update)

Kinetica.ConditionSet — Method

conditions = ConditionSet(Dict(
    :C1 => ConditionType1(...),
    :C2 => ConditionType2(...))
    [, ts_update]
))

Outer constructor for ConditionSet..

Separates condition profiles from their symbols and parses numeric profiles into StaticConditionProfiles. Registers AbstractVariableProfiles with Symbolics to allow for proper computation down the chain.

If ts_update is provided, creates tstops arrays within each variable profile for use in discrete rate update simulations.

Kinetica.get_profile — Method

profile = get_profile(cs, sym)

Gets the condition profile linked to sym from cs.

Kinetica.get_initial_conditions — Method

initial_conditions = get_initial_conditions(conditions)

Extract initial conditions from ConditionSet.

Returns an array of Pairs linking Symbols to initial values. For AbstractStaticProfiles, initial values are their static values. For AbstractVariableProfiles, initial values are their X_start values.

Kinetica.isstatic — Function

isstatic(cs[, sym])

Determines if condition profiles in a ConditionSet are static.

When a Symbol sym is provided, only checks if the profile linked to this Symbol is static. If no Symbol is provided, checks is all profiles are static.

isstatic(profile)

Determines if a given condition profile is static.

Kinetica.isvariable — Function

isvariable(cs[, sym])

Determines if condition profiles in a ConditionSet are variable.

When a Symbol sym is provided, only checks if the profile linked to this Symbol is variable. If no Symbol is provided, checks is all profiles are variable.

isvariable(profile)

Determines if a given condition profile is variable.

Kinetica.get_tstops — Method

tstops = get_tstops(cs)

Retrieves a sorted array of unique time stops from all condition profiles in cs.

Should be used for passing a unified set of time stops to a discrete rate constant update solver. Will throw an error if all condition profiles are static, as they have no tstops.

Kinetica.get_t_final — Method

t_final = get_t_final(cs)

Retrieves the last necessary time point needed to encompass all variable condition profiles in cs.

Will throw an error if all condition profiles are static, they have no set endpoints.

Kinetica.solve_variable_conditions! — Function

solve_variable_conditions!(cs, pars[, reset, solver, solve_kwargs])

Solves all variable condition profiles over the timespan in pars.tspan.

Places all condition profile solutions in their sol field. In the case of AbstractDirectProfiles, this creates a DiffEqArray to mimic the regular DiffEq solver interface.

Static Condition Profiles

Kinetica.StaticConditionProfile — Type

Directly Variable Condition Profiles

Kinetica.solve_variable_condition! — Method

solve_variable_condition!(profile<:AbstractDirectProfile, pars[, reset, solver, solve_kwargs])

Generates a solution for the specified directly-variable condition profile.

For profiles with direct functions, this requires calculating values for the specified pars.tspan and wrapping them within a DiffEqArray for compatibility with other interfaces (plotting, interpolation, etc.).

Arguments solver and solve_kwargs are provided for compatibility with callers, do nothing and should be ignored.

Kinetica.NullDirectProfile — Type

Container for null direct profile data and condition function.

This condition profile should only be used for debugging, as it has a condition function which always returns the initial condition. If only this constant condition is required, the regular ODESolve should always be used instead of an ODEConditionSolve with this condition profile.

Contains fields for:

Condition function (f)

Initial value of condition (X_start)

Time to stop calculation (t_end)

Times for the ODE solver to ensure calculation at (tstops)

Profile solution, constructed by call to solve_variable_condition! (sol)

Kinetica.NullDirectProfile — Method

condition_profile = NullDirectProfile(; X, t_end)

Outer constructor for null condition direct profile.

Should only be used for testing purposes (see struct documentation).

Kinetica.LinearDirectProfile — Type

Container for linear condition ramp profile data and condition function.

This condition profile represents a linear condition increase/decrease from X_start to X_end.

Contains fields for:

Condition function (f)

Rate of change of condition (rate)

Initial value of condition (X_start)

Final value of condition (X_end)

Time to stop calculation (t_end)

Times for the ODE solver to ensure calculation at (tstops)

Profile solution, constructed by call to solve_variable_condition! (sol)

Kinetica.LinearDirectProfile — Method

condition_profile = LinearDirectProfile(; rate, X_start, X_end)

Outer constructor for linear condition ramp direct profile.

Determines the simulation end time from the provided conditions and rate, then constructs the condition function (which is a linear y = mx + c function).

Gradient-Variable Condition Profiles

Kinetica.LinearGradientProfile — Type

Container for linear condition ramp profile data and condition gradient function.

This condition profile represents a linear condition increase/decrease from X_start to X_end.

Contains fields for:

Condition gradient function (grad)

Rate of change of condition (rate)

Initial value of condition (X_start)

Final value of condition (X_end)

Time to stop calculation (t_end)

Times for the ODE solver to ensure calculation at (tstops)

Profile solution, constructed by call to solve_variable_condition! (sol)

Kinetica.LinearGradientProfile — Method

condition_profile = LinearGradientProfile(; rate, X_start, X_end)

Outer constructor for linear condition ramp gradient profile.

Determines the simulation end time from the provided conditions and gradient, then constructs the condition gradient function (which returns rate for every timestep).

Kinetica.DoubleRampGradientProfile — Type

Container for double condition ramp profile data and condition gradient function.

This condition profile represents two condition ramps with adjustable condition plateaus before, after and in between the ramps, i.e.

              ------   X_mid
      rate1  /      \
            /        \  rate2
X_start ----          \
                       ----- X_end

The profile starts at X_start and maintains that value for t_start_plateau. The condition then ramps with gradient rate1 to condition value X_mid. This value is maintained for t_mid_plateau. The condition then ramps with gradient rate2 to condition value X_end. This value is maintained for t_end_plateau until the calculated time t_end.

To smooth out gradient discontinuities, a blending time t_blend can be passed to linearly interploate between plateaus and ramps, forming a smooth function of time. Larger values of t_blend yield smoother functions but decrease accuracy of the ramps, so should be used carefully.

Kinetica.DoubleRampGradientProfile — Method

condition_profile = DoubleRampGradientProfile(; X_start, t_start_plateau, rate1, X_mid, t_mid_plateau, rate2, X_end, t_end_plateau[, t_blend])

Outer constructor for double condition ramp gradient profile.

Determines the simulation end time from the provided conditions and gradients, then constructs the condition gradient function (which returns rate for every timestep).

If t_blend is passed, constructs a smooth approximation of the otherwise discontinuous gradient function.