Kinetica.jl API

Open Babel

Conversion

Property Calculation

Kinetica.get_species_stats! — Function

get_species_stats!(sd::SpeciesData[, refresh=false])

Gets statistics about the species in sd.

Calculates values for the following useful statistics, which are placed in the species data cache at sd.cache with a key for each property:

Average COM-atom radius of species (:radii)

Molecular weight of species (:weights)

Only calculates statistics for species without defined values, unless refresh=true which causes all statistics for every species to be updated.