Kinetica.jl API

Analysis

Kinetica.ODESolveOutput — Type

Data container for output of network ODE solutions.

Binds together all data required for analysis of network ODE solution results. Used as an input for many of the automated plotting functions in KineticaCore.

Also used for results IO, can be completely deconstructed into a package-independent dictionary tree for saving as binary JSON (BSON) with save_output. Can be reconstructed from such a BSON file with load_output.