Kinetica.jl API

CRN Representation

Representing Chemical Species

Kinetica.SpeciesData — Type

Bidirectional String-Int dictionary for chemical species.

Contains fields for:

SMILES string -> integer ID dictionary (toInt)

Integer ID -> SMILES string dictionary (toStr)

Number of species (n)

ExtXYZ structures of species (xyz)

Dictionary of per-species cached values (cache)

Base.push! — Method

push!(sd, smi, xyz)

Add a species to SpeciesData.

Does not account for smi already existing within sd. To ensure no overlap, use push_unique!.

Base.push! — Method

push!(sd, smis, xyzs)

Add an array of species to SpeciesData.

Does not account for smi already existing within sd. To ensure no overlap, use push_unique!.

Base.push! — Method

push!(sd, xyz_file[, fix_radicals])

Add all species in xyz_file to sd.

Does not account for smi already existing within sd. To ensure no overlap, use push_unique!.

Kinetica.push_unique! — Function

push_unique!(sd, smi, xyz)

Add a species SMILES to a SpeciesData, as long as it does not already exist there.

push_unique!(sd, xyz_file[, fix_radicals])

Add species in xyz_file to sd, as long as they do not already exist there.

push_unique!(sd, smis, xyzs)

Add an array of species to SpeciesData, as long as each does not already exist there.

Representing Reactions

Kinetica.RxData — Type

Data container for reactions.

Should be constructed alongside a SpeciesData for mapping species IDs to SMILES strings.

Contains fields for:

Number of reactions encountered (nr)

Atom-mapped reaction SMILES for unambiguous linking of atom indices in reactants and products (mapped_rxns)

Unique IDs of reactants for each reaction (id_reacs)

Unique IDs of products for each reaction (id_prods)

Stoichiometries of reactants for each reaction (stoic_reacs)

Stoichiometries of products for each reaction (stoic_prods)

Reaction enthalpies (dH)

Reaction hashes, used for unique identification (rhash)

Base.push! — Method

push!(rd, sd, reacs, prods, rsys, psys, dH[, unique_rxns, max_molecularity])

Base.splice! — Method

splice!(rd, rids)

Removes reactions at indeces rids from rd.

CRN Initialisation

Kinetica.init_network — Function

sd, rd = init_network([iType, fType])

Initialises an empty reaction network.

CDE Interface

Kinetica.CDE — Type

cde = CDE(rdir, template_dir, init_xyz, env[, sampling_seed, radius])

CDE runner. Initialised through struct, run through functor.

Struct contains fields for:

CDE template directory (template_dir)

Environmental multithreading number of threads (Optional, defaults to 1 thread; env_threads)

Path to CDE executable (Optional, defaults to CDE packaged within CDEjll; `cdeexec`)

Seed for CDE's RNG (Optional, setting to 0 indicates seed should be random; sampling_seed)

Radius for exploration of breakdown space (Optional, Default = 50; radius)

Number of mechanisms to generate within a single CDE run (Optional, Default = 1; nrxn)

Number of parallel CDE runs to execute (Optional, Default = 1; parallel_runs)

Maximum number of parallel CDE executables to run at any time (Optional, Default = 1; parallel_exes)

Whether to write CDE's stdout to file (Optional, Default = false; write_stdout)

Whether to write CDE's stderr to file (Optional, Default = false; write_stderr)

Whether to allow functions to continue running if CDE errors are detected (Optional, Default = false; allow_errors)

Additionally, some fields are usually modified within Kinetica, and are not intended to be changed by users.

Main reaction directory (rdir)

XYZ file of starting molecule/material (init_xyz)

Kinetica.ingest_cde_run — Function

reac_smis, reac_xyzs, reac_systems, prod_smis, prod_xyzs, prod_systems, dH = ingest_cde_run(rdir, rcount[, fix_radicals])

Reads in the results from a CDE run.

Separates out fragment species from each available reaction's reactants and products, forming arrays of their SMILES strings and ExtXYZ geometries.

OBCanonicalRadicals can be enabled to tidy up radical SMILES using the fix_radicals parameter.

Exploration

Kinetica.DirectExplore — Type

Container for parameters used in direct CRN exploration.

Contains fields for:

Top level of CRN exploration directory (rdir_head)

SMILES string(s) of main breakdown reactant(s) being studied (reac_smiles)

CDE instance (cde)

Maximum number of iterations to perform (maxiters)

Number of iterations with no change in reactions to consider as converged (rxn_convergence_threshold)

Kinetica.IterativeExplore — Type

Container for parameters used in iterative kinetics-based CRN exploration.

Contains fields for:

Top level of CRN exploration directory (rdir_head)

SMILES string(s) of main breakdown reactant(s) being studied (reac_smiles)

CDE instance (cde)

Maximum number of iterations to perform (maxiters)

Number of subspace iterations with no change in reactions to consider a subspace converged (rxn_convergence_threshold)

Number of level iterations with no change in seeds to consider the network converged (seed_convergence_threshold)

Concentration above which species will be selected as seeds each level (seed_conc)

Blacklist of species to avoid doing independent subspace explorations on (independent_blacklist)

Inert species that should not be considered for reaction (inert_species)

Kinetica.explore_network — Function

res = explore_network(exploremethod::Union{DirectExplore, IterativeExplore}, solvemethod[, savedir])

Runs network exploration with one of the available methods.

If exploremethod isa DirectExplore, runs a single-level network exploration to attempt to locate all relevant reactions in a radius of exploremethod.cde.radius species from the starting system.

If exploremethod isa IterativeExplore, runs a multi-level iterative network exploration using kinetic simulations to identify seed species for successive levels, in order to fully characterise the reaction space relevant to the conditions in solvemethod.conditions.

Molecule System

Kinetica.system_from_smiles — Function

system_from_smiles(smiles[, saveto, dmin])

Forms a single XYZ system out of the molecules in smiles.

Useful for making unified molecular systems with no overlap for feeding into CDE. dmin represents the minimum molecule-molecule distance that should be allowed.

If the argument saveto is provided, outputs the optimised system to a file at this path. If not, returns the optimised system as a single ExtXYZ dict.

Kinetica.system_from_mols — Function

Kinetica.molsys_opt — Function

tmols = molsys_opt(mols, dmin, maxiters)

Optimises positions of molecules in mols to ensure they are all at least dmin Angstroms apart.

Creates and solves an N-body spring-driven particle system and transforms molecular coordinates to these particles to check for proximity.